% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getNewDictionary}
\alias{getNewDictionary}
\title{Creates a .rds with the latest file.}
\usage{
getNewDictionary()
}
\value{
RDS is saved. If all went well, "OK" is returned
}
\description{
Connects to svn.code.sf.net to grab that
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
