% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{tonePlayer}
\alias{tonePlayer}
\title{Plays a string of syllables}
\usage{
tonePlayer(structure)
}
\arguments{
\item{structure}{string of 0,1,2, the output from `getStructure()`}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
WIP, tryng to figure out how to "sing" the inflection of a
sentance suitable for electronic music NLG
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_match_all}}
 \code{\link[Rmusic]{play_music}}
}
